+++ Running/src/main/java/runday/domain/RunningStarted.java
@@ -14,8 +14,8 @@ public class RunningStarted extends AbstractEvent {
     private Long id;
     private String userId;
     private Date startTime;
-    private Long startLatitude;
-    private Long startLongitude;
+    private Float startLatitude;
+    private Float startLongitude;
 
     public RunningStarted(Running aggregate) {
         super(aggregate);
